Fix enum generator analyzer failures in CI
